fix: move all test helper modules in test_tools.helpers
Address PR comments
