+++ autonomy/test_tools/configurations.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # ------------------------------------------------------------------------------
 #
-#   Copyright 2022 Valory AG
+#   Copyright 2021-2022 Valory AG
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -141,3 +141,4 @@ KEY_PAIRS: List[Tuple[str, str]] = [
         "0xdf57089febbacf7ba0bc227dafbffa9fc08a93fdc68e1e42411a14efcf23656e",
     ),
 ]
+ANY_ADDRESS = "0.0.0.0"  # nosec
