+++ scripts/whitelist.py
@@ -104,3 +104,13 @@ _._start_acn  # unused method (autonomy/test_tools/fixture_helpers.py:170)
 _._acn_node_image  # unused attribute (autonomy/test_tools/fixture_helpers.py:174)
 get_register_contract  # unused function (autonomy/test_tools/helpers/contracts.py:30)
 GanacheForkDockerImage  # unused function (autonomy/test_tools/docker/ganache.py:111)
+try_send  # unused function (/home/marcofavorito/workfolder/open-autonomy/autonomy/test_tools/helpers/base.py:68)
+make_round_class  # unused function (/home/marcofavorito/workfolder/open-autonomy/autonomy/test_tools/helpers/base.py:79)
+ANY_ADDRESS  # unused variable (/home/marcofavorito/workfolder/open-autonomy/autonomy/test_tools/configurations.py:144)
+wait_for_condition  # unused function (/home/marcofavorito/workfolder/open-autonomy/autonomy/test_tools/helpers/async_utils.py:29)
+BaseThreadedAsyncLoop  # unused class (/home/marcofavorito/workfolder/open-autonomy/autonomy/test_tools/helpers/async_utils.py:146)
+thread  # unused variable (/home/marcofavorito/workfolder/open-autonomy/autonomy/test_tools/helpers/async_utils.py:151)
+_.setup  # unused method (/home/marcofavorito/workfolder/open-autonomy/autonomy/test_tools/helpers/async_utils.py:153)
+_.execute  # unused method (/home/marcofavorito/workfolder/open-autonomy/autonomy/test_tools/helpers/async_utils.py:158)
+_.teardown  # unused method (/home/marcofavorito/workfolder/open-autonomy/autonomy/test_tools/helpers/async_utils.py:163)
+_.get_command  # unused method (/home/marcofavorito/workfolder/open-autonomy/autonomy/test_tools/helpers/tendermint_utils.py:125)
