+++ tests/test_agents/test_abstract_abci.py
@@ -24,9 +24,8 @@ import warnings
 import pytest
 from aea.test_tools.test_cases import AEATestCaseMany
 
+from autonomy.test_tools.configurations import ANY_ADDRESS
 from autonomy.test_tools.fixture_helpers import UseTendermint
-
-from tests.conftest import ANY_ADDRESS
 
 
 @pytest.mark.e2e
