+++ tests/test_skills/test_abstract_round_abci/test_abci_app_chain.py
@@ -22,13 +22,13 @@
 import pytest
 from aea.exceptions import AEAEnforceError
 
+from autonomy.test_tools.helpers.base import make_round_class
+
 from packages.valory.skills.abstract_round_abci.abci_app_chain import (
     AbciAppTransitionMapping,
     chain,
 )
 from packages.valory.skills.abstract_round_abci.base import AbciApp, DegenerateRound
-
-from tests.helpers.base import make_round_class
 
 
 class TestAbciAppChaining:
